Move options menu and  toolbar from ScannerFragment to MainActivity
