+++ app/src/main/java/pl/pola_app/ui/activity/MainActivity.java
@@ -2,11 +2,17 @@ package pl.pola_app.ui.activity;
 
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
+import android.content.ActivityNotFoundException;
 import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.RecyclerView;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.widget.Toast;
 
 import com.squareup.otto.Bus;
@@ -20,6 +26,7 @@ import pl.pola_app.R;
 import pl.pola_app.helpers.EventLogger;
 import pl.pola_app.helpers.ProductsListLinearLayoutManager;
 import pl.pola_app.helpers.SessionId;
+import pl.pola_app.helpers.Utils;
 import pl.pola_app.model.SearchResult;
 import pl.pola_app.ui.adapter.ProductList;
 import pl.pola_app.ui.adapter.ProductsAdapter;
@@ -34,6 +41,8 @@ public class MainActivity extends AppCompatActivity implements MainViewBinder {
     Bus eventBus;
     @Bind(R.id.products_list)
     RecyclerView productsListView;
+    @Bind(R.id.toolbar)
+    Toolbar toolbar;
 
     private ScannerFragment scannerFragment;
     private MainPresenter mainPresenter;
@@ -58,6 +67,19 @@ public class MainActivity extends AppCompatActivity implements MainViewBinder {
         scannerFragment.setOnBarcodeScannedListener(mainPresenter);
 
         productsListView.setLayoutManager(new ProductsListLinearLayoutManager(this));
+
+        setupActionBar();
+    }
+
+    private void setupActionBar() {
+        setSupportActionBar(toolbar);
+        setTitle(getString(R.string.app_name));
+        final ActionBar supportActionBar = getSupportActionBar();
+        if (supportActionBar != null) {
+            supportActionBar.setTitle("");
+            supportActionBar.setDisplayHomeAsUpEnabled(false);
+            supportActionBar.setHomeButtonEnabled(false);
+        }
     }
 
     @Override
@@ -129,6 +151,78 @@ public class MainActivity extends AppCompatActivity implements MainViewBinder {
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         Nammu.onRequestPermissionsResult(requestCode, permissions, grantResults);
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // handle item selection
+        Intent intent;
+        switch (item.getItemId()) {
+            case R.id.action_about:
+                intent = new Intent(this, ActivityWebView.class);
+                intent.putExtra("url", Utils.URL_POLA_ABOUT);
+                startActivity(intent);
+                return true;
+            case R.id.action_metodology:
+                intent = new Intent(this, ActivityWebView.class);
+                intent.putExtra("url", Utils.URL_POLA_METHOD);
+                startActivity(intent);
+                return true;
+            case R.id.action_club:
+                intent = new Intent(this, ActivityWebView.class);
+                intent.putExtra("url", Utils.URL_POLA_KJ);
+                startActivity(intent);
+                return true;
+            case R.id.action_team:
+                intent = new Intent(this, ActivityWebView.class);
+                intent.putExtra("url", Utils.URL_POLA_TEAM);
+                startActivity(intent);
+                return true;
+            case R.id.action_partners:
+                intent = new Intent(this, ActivityWebView.class);
+                intent.putExtra("url", Utils.URL_POLA_PARTNERS);
+                startActivity(intent);
+                return true;
+            case R.id.action_bug:
+                intent = new Intent(this, CreateReportActivity.class);
+                intent.setAction("product_report");
+                startActivity(intent);
+                return true;
+            case R.id.action_mail:
+                Intent emailIntent = new Intent(Intent.ACTION_SENDTO, Uri.fromParts("mailto", Utils.POLA_MAIL, null));
+                emailIntent.putExtra(Intent.EXTRA_SUBJECT, "Pola");
+                startActivity(Intent.createChooser(emailIntent, getString(R.string.send_email_picker)));
+                return true;
+            case R.id.action_rate:
+                Uri uri = Uri.parse("market://details?id=" + this.getPackageName());
+                intent = new Intent(Intent.ACTION_VIEW, uri);
+                intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY |
+                        Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET |
+                        Intent.FLAG_ACTIVITY_MULTIPLE_TASK);
+                try {
+                    startActivity(intent);
+                } catch (ActivityNotFoundException e) {
+                    startActivity(new Intent(Intent.ACTION_VIEW,
+                            Uri.parse("http://play.google.com/store/apps/details?id=" + this.getPackageName())));
+                }
+                return true;
+            case R.id.action_fb:
+                intent = new Intent(Intent.ACTION_VIEW, Uri.parse(Utils.URL_POLA_FB));
+                startActivity(intent);
+                return true;
+            case R.id.action_twitter:
+                intent = new Intent(Intent.ACTION_VIEW, Uri.parse(Utils.URL_POLA_TWITTER));
+                startActivity(intent);
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
     }
 
 }
