+++ app/src/main/res/layout/activity_main.xml
@@ -11,10 +11,13 @@
         android:layout_height="match_parent"
         android:name="pl.pola_app.ui.fragment.ScannerFragment"
         android:id="@+id/scanner_fragment"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
         tools:layout="@layout/fragment_scanner" />
+
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="@android:color/transparent" />
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/products_list"
